INF: Glossary and Lab1 for gestione-dati module.
The Lab 01 "esplorazione dataset su file" for the 'gestione-dat' module
has been filled with first draft of the directions. The laboratory
consists of dataset parsing when these are stored in plain files (like
CSV and JSON).

We also added the glossary for the same module adding the terms (without
definitions for now) most important for the module. This represents a
novelty in the website, but it may become a standard part of all (or
most) modules, which may even deserve a dedicated template (possibly
with code for a more usable glossary page).

Also some changes have been made di the lecture one 'informazione-dati'
of the module.