Started 02-rappresentazione.md, new lecture files.
In 02-rappresentazione.md we added some of the key definitions that
must be acquired during the lecture as well as some correlated text.
The page is still incomplete and in a preliminary state.

We also added some new files that will host lecture content, updated
also the glossary of the unit.