INF: More on SQL DDL module.
On 01-ddl-create.mde added description for 'ON DELETE' constraint.

Completed first draft of the 03-ddl-update-delete.md and
04-ddl-alter-drop.md lectures describing the SQL instructions: UPDATE,
DELETE, ALTER, and DROP. All lectures have examples, but they still
lack reviewing questions and exercises.

Started laboratory 1 lecture on using the SQL DDL instructions to
create a database for an e-commerce website.