Entire refactoring of pages.
All parts of the site are now "dynamically" generate through routes
and generateStaticParams. Implemented the basics for all types of routes
- [subject]
- [subject]/[year]
- [subject]/[year]/[module]
- [subject]/[year]/[module]/[lecture]
Currently, all information is in the Subjects object within site.config,
in the future this object could be created by scanning configuration
files and, more generally, the file system with content.